% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204175R.R
\name{las.kA}
\alias{las.kA}
\title{Oracle Trans-Lasso}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{X}{target design matrix}

\item{y}{response}

\item{A0}{informative auxiliary samples}

\item{n.vec}{sample index}

\item{lam.const}{tuning parameter}

\item{l1}{l1=T means using l1 method}
}
\description{
Oracle Trans-Lasso
}
