% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204175R.R
\name{mse.fun}
\alias{mse.fun}
\title{computing the MSE}
\usage{
mse.fun(beta, est, X.test = NULL)
}
\arguments{
\item{beta}{coefficient vector}

\item{est}{estimator}

\item{X.test}{test dataset}
}
\description{
computing the MSE
}
