% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mse_fun}
\alias{mse_fun}
\title{computing the MSE}
\usage{
mse_fun(beta, est, X_test = NULL)
}
\arguments{
\item{beta}{coefficient vector}

\item{est}{estimator}

\item{X_test}{test dataset}
}
\description{
computing the MSE
}
